Migrate the existing configuration workflow from JSON to YAML (#40)
* feat(config): migrate the existing configuration workflow from JSON to YAML

The current configuration file, which is in JSON format, has become difficult to maintain. While it was sufficient during the beta phase, it is no longer suitable for our needs. This change to YAML is necessary to improve maintainability and readability.

Please note that there are no plans to build an automated converter from JSON to YAML. If an automated conversion is essential for your workflow, please create an issue on GitHub to discuss it further.

* ci: remove pull request from the ci workflow